requirements.txt
lxml==4.6.3
Pillow==8.2.0
python-pptx==0.6.18
XlsxWriter==1.3.8
